+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   (for [#1221](https://github.com/pbek/QOwnNotes/issues/1221)
 - there now is a new checkbox `Ignore code font size to allow zooming` in the
   *Preview settings* to allow zooming of code fonts (for [#1202](https://github.com/pbek/QOwnNotes/issues/1202)
+    - this checkbox is on by default
+- fixed search widget colors in the log widget in dark mode
 
 ## 19.6.4
 - if you right-click on a link in the preview you now can select `Copy link location`